+++ src/steps/containers/client.ts
@@ -2,7 +2,7 @@ import { google, container_v1 } from 'googleapis';
 import { Client } from '../../google-cloud/client';
 
 export class ContainerClient extends Client {
-  private client = google.container('v1');
+  private client = google.container({ version: 'v1', retry: false });
 
   async iterateClusters(
     callback: (data: container_v1.Schema$Cluster) => Promise<void>,